# artifacts/api-server/src/routes/auth.ts
import { Router, type IRouter } from "express";
import { randomBytes } from "crypto";
import { db, usersTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import { nanoid } from "nanoid";
import {
  RegisterBody,
  LoginBody,
  VerifyEmailBody,
  ResendVerificationBody,
  ChangePasswordBody,
  ForgotPasswordBody,
} from "@workspace/api-zod";
import {
  signToken,
  hashPassword,
  comparePassword,
  requireAuth,
  getUserById,
} from "../lib/auth";
import {
  sendVerificationEmail,
  sendPasswordResetEmail,
} from "../lib/email";

const router: IRouter = Router();

function formatUser(user: typeof usersTable.$inferSelect) {
  return {
    id: user.id,
    fullName: user.fullName,
    email: user.email,
    mobile: user.mobile,
    mobileMasked: user.mobile.replace(/(\+91\s?)(\d{2})\d{6}(\d{2})/, "$1$2******$3"),
    role: user.role,
    status: user.status,
    emailVerified: user.emailVerified,
    isApproved: user.status === "approved",
    photoUrl: null as string | null,
    createdAt: user.createdAt?.toISOString() ?? new Date().toISOString(),
  };
}

router.post("/auth/register", async (req, res): Promise<void> => {
  const parsed = RegisterBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { fullName, mobile, email, password } = parsed.data;

  const [existing] = await db.select().from(usersTable).where(eq(usersTable.email, email));
  if (existing) {
    res.status(400).json({ error: "Email already registered" });
    return;
  }

  const count = await db.select().from(usersTable);
  const isFirstUser = count.length === 0;

  const passwordHash = await hashPassword(password);
  const verifyToken = randomBytes(3).toString("hex").toUpperCase();
  const verifyExpiry = new Date(Date.now() + 24 * 60 * 60 * 1000);

  const [user] = await db.insert(usersTable).values({
    id: nanoid(),
    fullName,
    email,
    mobile,
    passwordHash,
    role: isFirstUser ? "super_admin" : "student",
    status: isFirstUser ? "approved" : "pending_approval",
    emailVerified: false,
    emailVerifyToken: verifyToken,
    emailVerifyExpiry: verifyExpiry,
  }).returning();

  await sendVerificationEmail(email, verifyToken);

  const token = signToken({ id: user.id, email: user.email, role: user.role, status: user.status });
  res.status(201).json({
    user: formatUser(user),
    token,
    message: "Registration successful. Please verify your email.",
  });
});

router.post("/auth/login", async (req, res): Promise<void> => {
  const parsed = LoginBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { email, password } = parsed.data;

  const [user] = await db.select().from(usersTable).where(eq(usersTable.email, email));
  if (!user) {
    res.status(401).json({ error: "Invalid credentials" });
    return;
  }

  const valid = await comparePassword(password, user.passwordHash);
  if (!valid) {
    res.status(401).json({ error: "Invalid credentials" });
    return;
  }

  if (!user.emailVerified) {
    res.status(403).json({ error: "Please verify your email before logging in" });
    return;
  }

  if (user.status === "suspended" || user.status === "banned") {
    res.status(403).json({ error: "Account suspended or banned" });
    return;
  }

  await db.update(usersTable).set({ lastLoginAt: new Date() }).where(eq(usersTable.id, user.id));

  const token = signToken({ id: user.id, email: user.email, role: user.role, status: user.status });
  res.json({ user: formatUser(user), token });
});

router.post("/auth/logout", async (_req, res): Promise<void> => {
  res.json({ success: true, message: "Logged out" });
});

router.post("/auth/verify-email", async (req, res): Promise<void> => {
  const parsed = VerifyEmailBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const { token } = parsed.data;
  const [user] = await db.select().from(usersTable).where(eq(usersTable.emailVerifyToken, token));
  if (!user) {
    res.status(400).json({ error: "Invalid or expired token" });
    return;
  }

  if (user.emailVerifyExpiry && user.emailVerifyExpiry < new Date()) {
    res.status(400).json({ error: "Token expired" });
    return;
  }

  await db.update(usersTable)
    .set({ emailVerified: true, emailVerifyToken: null, emailVerifyExpiry: null })
    .where(eq(usersTable.id, user.id));

  res.json({ success: true, message: "Email verified successfully" });
});

router.post("/auth/resend-verification", async (req, res): Promise<void> => {
  const parsed = ResendVerificationBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [user] = await db.select().from(usersTable).where(eq(usersTable.email, parsed.data.email));
  if (!user || user.emailVerified) {
    res.json({ success: true, message: "If that email is registered and unverified, a new code was sent" });
    return;
  }

  const verifyToken = randomBytes(3).toString("hex").toUpperCase();
  const verifyExpiry = new Date(Date.now() + 24 * 60 * 60 * 1000);

  await db.update(usersTable)
    .set({ emailVerifyToken: verifyToken, emailVerifyExpiry: verifyExpiry })
    .where(eq(usersTable.id, user.id));

  await sendVerificationEmail(user.email, verifyToken);
  res.json({ success: true, message: "Verification email sent" });
});

router.post("/auth/change-password", requireAuth, async (req, res): Promise<void> => {
  const parsed = ChangePasswordBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const user = await getUserById(req.user!.id);
  if (!user) {
    res.status(404).json({ error: "User not found" });
    return;
  }

  const valid = await comparePassword(parsed.data.currentPassword, user.passwordHash);
  if (!valid) {
    res.status(400).json({ error: "Current password is incorrect" });
    return;
  }

  const hash = await hashPassword(parsed.data.newPassword);
  await db.update(usersTable).set({ passwordHash: hash }).where(eq(usersTable.id, user.id));

  res.json({ success: true, message: "Password changed successfully" });
});

router.post("/auth/forgot-password", async (req, res): Promise<void> => {
  const parsed = ForgotPasswordBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [user] = await db.select().from(usersTable).where(eq(usersTable.email, parsed.data.email));
  if (user) {
    const resetToken = randomBytes(3).toString("hex").toUpperCase();
    const resetExpiry = new Date(Date.now() + 60 * 60 * 1000);
    await db.update(usersTable)
      .set({ passwordResetToken: resetToken, passwordResetExpiry: resetExpiry })
      .where(eq(usersTable.id, user.id));
    await sendPasswordResetEmail(user.email, resetToken);
  }

  res.json({ success: true, message: "If that email is registered, a reset link was sent" });
});

router.get("/auth/me", requireAuth, async (req, res): Promise<void> => {
  const user = await getUserById(req.user!.id);
  if (!user) {
    res.status(404).json({ error: "User not found" });
    return;
  }
  res.json(formatUser(user));
});

export { formatUser };
export default router;