# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdTech Study Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
        "403":
          description: Email not verified
...
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [auth]
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationInput"
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password (requires auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated

  # Profile
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_approval, approved, suspended, banned]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}/approve:
    post:
      operationId: approveUser
      tags: [admin]
      summary: Approve a pending user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Change user role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # Auth schemas
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        mobile:
          type: string
          description: "Indian mobile number format: +91XXXXXXXXXX"
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - fullName
        - mobile
        - email
        - password
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    VerifyEmailInput:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    ResendVerificationInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ChangePasswordInput:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
        message:
          type: string
      required:
        - user
        - token
    UserProfile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: string
        mobileMasked:
          type: string
        role:
          type: string
          enum: [super_admin, admin, student]
        status:
          type: string
          enum: [pending_approval, approved, suspended, banned]
        emailVerified:
          type: boolean
        isApproved:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - mobile
        - mobileMasked
        - role
        - status
        - emailVerified
        - isApproved
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [super_admin, admin, student]
        status:
          type: string
          enum: [pending_approval, approved, suspended, banned]
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
      required:
        - id
        - fullName
        - email
        - mobile
        - role
        - status
        - emailVerified
        - createdAt
    RoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, student]
      required:
        - role
...